[dashboard] move dashboard/public/share folder into dashboard/public/dashboard_app/top_nav/share
