+++ src/platform/plugins/shared/dashboard/public/dashboard_app/top_nav/share/export_json/export_json_config.tsx
