+++ src/platform/plugins/shared/dashboard/public/dashboard_app/top_nav/share/export_json/sanitize_dashboard.ts
@@ -7,9 +7,9 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import type { DashboardSanitizeResponseBody, DashboardState } from '../../server';
-import { DASHBOARD_INTERNAL_API_PATH } from '../../common/constants';
-import { coreServices } from '../services/kibana_services';
+import type { DashboardSanitizeResponseBody, DashboardState } from '../../../../../server';
+import { DASHBOARD_INTERNAL_API_PATH } from '../../../../../common/constants';
+import { coreServices } from '../../../../services/kibana_services';
 
 export async function sanitizeDashboard(dashboardState: DashboardState) {
   const result = await coreServices.http.post<DashboardSanitizeResponseBody>(
