+++ src/platform/plugins/shared/dashboard/public/dashboard_renderer/dashboard_module.ts
@@ -19,5 +19,5 @@ export { UnlinkFromLibraryAction } from '../dashboard_actions/library_unlink_act
 export { CopyToDashboardAction } from '../dashboard_actions/copy_to_dashboard_action';
 export { AddSectionAction } from '../dashboard_actions/add_section_action';
 export { dashboardDrilldown } from '../dashboard_drilldown/dashboard_drilldown';
-export { exportJsonConfig } from '../share/export_json_config';
+export { exportJsonConfig } from '../dashboard_app/top_nav/share/export_json/export_json_config';
 export { initializeDashboardApiServices } from '../services/dashboard_api_services';
